Add the possibility to have variable size outputs (#5823)
Summary:
Add the possibility to specify a node with a variable number of outputs (we already had this feature for inputs but not for outputs)

### Documentation:
This feature is useful in some cases we have encountered.

The most visible change relates to the node API change : Node::getOutputName() now returns a std::string instead of a llvm::StringRef. This gives the freedom to generate output names on the fly and makes the output API consistent with the input API (that already returns  std::string for the same reason).

Example:
_BB.newBackendSpecificNode("TestNodeVector")
    .addMember(MemberType::VectorNodeValue, "Inputs")
    .addVariableResult("Outputs")
    .setDocstring("Node with variable number of inputs and outputs,")_;

Pull Request resolved: https://github.com/pytorch/glow/pull/5823

Test Plan: This is tested internally since we use this feature. I didn't add a unit test in this PR because this would mean adding a fake node to Glow just for testing purpose. Another solution is to create a backend just for testing this, what seems overkill.

Reviewed By: khabinov

Differential Revision: D36137834

Pulled By: opti-mix

fbshipit-source-id: d9e54d830f1db5808df3b0c8295435d878c7b0d1