Support for multiple user-defined memory regions (#5956)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5956

This diff adds a feature that allows for very flexible and configurable definition of memory regions. It is a generalization of the current Glow mechanism, which has only three default memory regions: Mutable Weights Region for all inputs/outputs, Constant Weights Region and Activation Region for activations and intermediate results.

Memory regions can be used to group different tensors together in the same region of memory and specify various properties for this region of memory, for example:
- Which tensors should be in this region of memory? E.g. only FC weights, or only input tensors. One can define region matching rules for tensors using predicates and AND/OR conditions to specify which tensors belong to each memory region
- Which memory allocation algorithm to use?
- Can tensors in this region reuse memory or not?
- Should each tensor matching the rule be in its own memory region or should they be combined?
- Which part of the memory hierarchy should be used for a memory region, e.g. DDR or faster memories?

Descriptions of memory regions could be potentially defined programmatically or e.g. by means of a config file.

The feature is useful for performance experiments and deployment experiments. E.g. selectively placing specific kinds of tensors at best-suited levels of the memory hierarchy can significantly improve performance.

Memory allocation and memory management mechanisms of the compiler are reformulated to be memory region based. If no user-defined memory regions definitions are provided by the user, default memory regions definitions are used, which correspond to the old compiler behaviour and are supposed to be backwards compatible.

Reviewed By: jfix71, yuhc

Differential Revision: D33467780

fbshipit-source-id: cf283073a4591fac9856e7082df97d3fdd35d239