Handle removal of torch.split wrapper
Summary:
A recent change https://github.com/pytorch/pytorch/pull/74727 was reverted due to test failures, but needs to be re-landed.

The issue with tests was that we're expecting `torch.split` to be JIT traced into `aten::split_with_sizes`, but this won't be the case after the original diff is re-landed.

Reviewed By: mehtanirav

Differential Revision: D37766943

fbshipit-source-id: c6a0fc408fa8462ab405fea9f50c712915383298