Handle removal of torch.split wrapper (#5975)
Summary:
Pull Request resolved: https://github.com/pytorch/glow/pull/5975

A recent change https://github.com/pytorch/pytorch/pull/74727 was reverted due to test failures, but needs to be re-landed.

The issue with tests was that we're expecting `torch.split` to be JIT traced into `aten::split_with_sizes`, but this won't be the case after the original diff is re-landed.

Reviewed By: mehtanirav

Differential Revision: D37766943

fbshipit-source-id: 2b5b8f28398d14a4e393f2109d1ed93833c1d589