extend tbe benchmark to support external data in the same format as tbe on A100.
Summary:
Add an command line argument to load external data in the TBE benchmark.
The input file are pytorch tensors stored by the code:
torch.save([indices, offsets, lengths], path)
indices and offsets are the stacked sparse matrix in CSR format.
length is a tensor of size (number of tables) * (batch size). It is used only to determine the two parameters of the request.
When the external data path is given, the number of tables and batch size is determined by the input file and their setting in command line args will be ignored.
Add -data_file command line option in run_bench.py script to invoke TBE benchmark with external data.

Reviewed By: mdschatz, jiyuanzFB

Differential Revision: D38646117

fbshipit-source-id: d8577359ef78f4fde65d181c8897cb26ef813c21