Remove CI (#6136)
Summary:
Discussion seems to indicate that the CI running in OSS for this repo is no longer useful. Since we are moving away from CircleCI to GHA, we should deprecate these CI jobs, and they can be brought back in GitHub Actions at a later time if desired. This PR removes the CircleCI config, and CI jobs will stop running in OSS.

Based on the discussions internally, openrichardfb, jfix71, jackm321, and Huamin seem to be on-board.

Pull Request resolved: https://github.com/pytorch/glow/pull/6136

Reviewed By: osalpekar

Differential Revision: D49709373

Pulled By: openrichardfb

fbshipit-source-id: a2712adca6e5c9b51caee54a472eeeb60c1cb543